+++ lib/joystick/joystick_controller.dart
@@ -1,4 +1,5 @@
 import 'package:bonfire/base/bonfire_game.dart';
+import 'package:bonfire/joystick/joystick.dart';
 import 'package:bonfire/util/bonfire_game_ref.dart';
 import 'package:bonfire/util/mixins/pointer_detector.dart';
 import 'package:bonfire/util/priority_layer.dart';
@@ -59,7 +60,8 @@ mixin JoystickListener {
 abstract class JoystickController extends Component
     with BonfireHasGameRef<BonfireGame>, PointerDetectorHandler {
   List<JoystickListener> _observers = [];
-  bool keyboardEnable = false;
+
+  KeyboardConfig keyboardConfig = KeyboardConfig(enable: false);
 
   void onKeyboard(RawKeyEvent event) {}