+++ lib/objects/animated_follower_object.dart
@@ -36,7 +36,7 @@ class AnimatedFollowerObject extends FollowerObject {
     super.update(dt);
     if (!loopAnimation) {
       if (animation?.isLastFrame == true) {
-        remove();
+        removeFromParent();
       }
     }
   }