+++ lib/widgets/bonfire_tiled_widget.dart
@@ -104,8 +104,8 @@ class _BonfireTiledWidgetState extends State<BonfireTiledWidget>
     if (widget.constructionMode) {
       widget.map.build().then((value) async {
         await _game?.map.updateTiles(value.map.tiles);
-        _game?.decorations().forEach((d) => d.remove());
-        _game?.enemies().forEach((e) => e.remove());
+        _game?.decorations().forEach((d) => d.removeFromParent());
+        _game?.enemies().forEach((e) => e.removeFromParent());
         value.components?.forEach((d) => _game?.addGameComponent(d));
       });
     }