+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# [next]
+- update flame to `1.0.0-releasecandidate.15`
+- [BREAKING CHANGE] now use `removeFromParent` instead of `remove` to remove the component of the game.
+- [BREAKING CHANGE] remove param `keyboardEnable` and `keyboardDirectionalType` in `Joystick`. Move this attributes to `keyboardConfig`
+
 # 1.8.1
 - Adds `Focus` in `CustomGameWidget` to remove "system ding" in MACOS.
 - Updates `moveToTarget` method in `Camera` to receive null;